Clarifying how mtimes are stored and compared.
The `math.isclose()` comparison was failing because even a 1s
difference was _numerically_ the same mtime.

Although the mtime read from disk would be consistent across multiple
runs, the next issue was that reading in a string from the database and
then casting it to a float will result in a different number from the
original, and we have to be careful about the string representation out
to a certain number of digits.

This commit clarifies that we're formatting time strings out to 6
digits, and this is now consistent across `taskgraph`.

RE:#70