Using nanosecond mtime resolution.
The various approaches I've tried so far in this work have been focusing
around comparing floats, which is the source of the problem anyways.
Ints should be easier to represent in the database, easier and more
reliable to cast from strings and easier to ultimately compare for when
there's a change.  RE:#70