Failing test now relies on task.join() rather than graph.join()
The failure in question was happening on linux because the graph can get
in to a state where the task has failed but we haven't yet reached
join().  In this state, the graph may have been terminated by an
exception, but that exception, while logged, is not raised.

An intermediate fix for this is to simply move the exception assertion
to contain task.join(), where the exception will be raised.  But the
longer-term solution here may well be to re-raise exceptions
from TaskGraph.join(). RE:#76